refactor: modularize health monitoring core
